# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clinical Trial Data Collection Platform API
  description: |
    HIPAA-compliant RESTful API for clinical trial patient data collection platform.
    
    This API provides secure endpoints for:
    - User authentication with JWT and MFA support
    - Patient data management with column-level encryption
    - Questionnaire creation and assignment
    - Secure response collection and storage
    - Media file uploads with virus scanning
    - Comprehensive audit logging
    
    ## Security Features
    - JWT Bearer token authentication
    - Role-based access control (Physician vs Patient)
    - Multi-factor authentication (MFA) support
    - Column-level data encryption for PHI
    - HIPAA-compliant audit logging
    - Rate limiting and request throttling
    - Virus scanning for file uploads
    
    ## Data Privacy
    - All personally identifiable information (PII) is encrypted at rest
    - PHI is never exposed in error messages
    - Comprehensive audit trails for all data access
    - Secure token-based session management
    
    ## Authentication Flow
    1. Login with email/password to receive access token
    2. Include Bearer token in Authorization header for all requests
    3. Refresh tokens before expiration using /auth/refresh
    4. Complete MFA verification if enabled
    
  version: 1.0.0
  contact:
    name: Clinical Trial Platform Development Team
    email: dev@clinicaltrial.platform
  license:
    name: Proprietary
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token obtained from /auth/login endpoint.
        
        Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  schemas:
...
    Questionnaire:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique questionnaire identifier
        title:
          type: string
          description: Questionnaire title
          example: Daily Symptom Assessment
        description:
          type: string
          description: Detailed description of the questionnaire
          example: Track daily symptoms and medication effectiveness
        version:
          type: string
          description: Questionnaire version
          example: "1.2"
        type:
          type: string
          enum: [symptom_tracker, medication_adherence, quality_of_life, side_effects, custom]
          description: Type of questionnaire
          example: symptom_tracker
        frequency:
          type: string
          enum: [once, daily, weekly, monthly, as_needed]
          description: How often the questionnaire should be completed
          example: daily
        isRequired:
          type: boolean
          description: Whether completion is mandatory
          example: true
        estimatedTimeMinutes:
          type: integer
          minimum: 1
          maximum: 120
          description: Estimated completion time in minutes
          example: 10
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
          description: Array of questions in the questionnaire
        status:
          type: string
          enum: [draft, active, archived]
          description: Current questionnaire status
          example: active
        publishedAt:
          type: string
          format: date-time
          description: When the questionnaire was published
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - title
        - version
        - type
        - frequency
        - status
    Question:
      type: object
      properties:
        id:
          type: string
          description: Unique question identifier
        type:
          type: string
          enum: [text, number, scale, multiple_choice, single_choice, date, time, boolean]
          description: Type of question
          example: scale
        text:
          type: string
          description: Question text
          example: How would you rate your pain level today?
        required:
          type: boolean
          description: Whether this question is required
          example: true
        options:
          type: array
          items:
            type: string
          description: Available options for choice questions
          example: ["No pain", "Mild pain", "Moderate pain", "Severe pain"]
        validation:
          type: object
          description: Validation rules for the question
          properties:
            min:
              type: number
              description: Minimum value for numeric/scale questions
            max:
              type: number
              description: Maximum value for numeric/scale questions
            minLength:
              type: integer
              description: Minimum text length
            maxLength:
              type: integer
              description: Maximum text length
      required:
        - id
        - type
        - text
        - required
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique response identifier
        questionnaireId:
          type: string
          format: uuid
          description: Associated questionnaire ID
        questionnaireVersion:
          type: string
          description: Version of questionnaire used
          example: "1.2"
        status:
          type: string
          enum: [not_started, in_progress, completed, submitted, reviewed]
          description: Current status of the response
          example: completed
        startedAt:
          type: string
          format: date-time
          description: When the response was started
        completedAt:
          type: string
          format: date-time
          description: When the response was completed
        submittedAt:
          type: string
          format: date-time
          description: When the response was submitted
        reviewedAt:
          type: string
          format: date-time
          description: When the response was reviewed by physician
        autoSavedAt:
          type: string
          format: date-time
          description: Last auto-save timestamp
        responseData:
          type: object
          description: Encrypted response data (only returned to authorized users)
          additionalProperties: true
      required:
        - id
        - questionnaireId
        - questionnaireVersion
        - status
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Authentication required
        code:
          type: string
          description: Machine-readable error code
          example: AUTH_REQUIRED
        details:
          type: array
          items:
            type: object
          description: Additional error details (validation errors, etc.)
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        requestId:
          type: string
          description: Unique request identifier for debugging
      required:
        - error
        - code
    PaginationResponse:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
          example: 10
        total:
          type: integer
          minimum: 0
          description: Total number of items
          example: 25
        pages:
          type: integer
          minimum: 1
          description: Total number of pages
          example: 3
      required:
        - page
        - limit
        - total
        - pages
paths:
...
  /patients/questionnaires/{questionnaireId}:
    get:
      tags:
        - Patients
      summary: Get questionnaire details with patient's response
      description: |
        Retrieve detailed questionnaire information including questions and any existing
        patient response data.
        
        Response data is decrypted and returned if the patient has started the questionnaire.
        Access permissions are verified before returning data.
      parameters:
        - name: questionnaireId
          in: path
          required: true
          description: Unique questionnaire identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Questionnaire details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  questionnaire:
                    allOf:
                      - $ref: '#/components/schemas/Questionnaire'
                      - type: object
                        properties:
                          response:
                            allOf:
                              - $ref: '#/components/schemas/PatientResponse'
                              - type: object
                                description: Current patient response (null if not started)
        '403':
          description: Access denied to questionnaire
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                error: Access denied
                message: You do not have access to this questionnaire
                code: QUESTIONNAIRE_ACCESS_DENIED
        '404':
          description: Questionnaire not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /questionnaires:
    get:
      tags:
        - Questionnaires
      summary: Get questionnaires (filtered by user type)
      description: |
        Retrieve questionnaires based on user permissions:
        - Patients: Only questionnaires they have access to
        - Physicians: Questionnaires they created or have access to through studies
        
        Results are paginated and can be filtered by type, status, and search terms.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by questionnaire status
          schema:
            type: string
            enum: [draft, active, archived]
            default: active
        - name: type
          in: query
          description: Filter by questionnaire type
          schema:
            type: string
            enum: [symptom_tracker, medication_adherence, quality_of_life, side_effects, custom]
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: Questionnaires retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  questionnaires:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Questionnaire'
                        - type: object
                          properties:
                            creator:
                              type: string
                              description: Name of the physician who created the questionnaire
                            studyTitle:
                              type: string
                              description: Associated study name
                            stats:
                              type: object
                              properties:
                                totalResponses:
                                  type: integer
                                completedResponses:
                                  type: integer
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
    post:
      tags:
        - Questionnaires
      summary: Create new questionnaire (physicians only)
      description: |
        Create a new questionnaire with questions and validation rules.
        
        Only physicians can create questionnaires. The questionnaire is created in
        'draft' status and must be explicitly published to become available to patients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Questionnaire title
                description:
                  type: string
                  maxLength: 1000
                  description: Detailed description
                version:
                  type: string
                  default: "1.0"
                  description: Version number
                studyId:
                  type: string
                  format: uuid
                  description: Associated study ID (optional)
                questionnaireType:
                  type: string
                  enum: [symptom_tracker, medication_adherence, quality_of_life, side_effects, custom]
                frequency:
                  type: string
                  enum: [once, daily, weekly, monthly, as_needed]
                scheduleData:
                  type: object
                  description: Scheduling configuration for recurring questionnaires
                questions:
                  type: array
                  items:
                    $ref: '#/components/schemas/Question'
                  minItems: 1
                validationRules:
                  type: object
                  description: Global validation rules for the questionnaire
                isRequired:
                  type: boolean
                  default: false
                estimatedTimeMinutes:
                  type: integer
                  minimum: 1
                  maximum: 120
                instructions:
                  type: string
                  maxLength: 2000
              required:
                - title
                - questionnaireType
                - frequency
                - questions
            example:
              title: Daily Pain Assessment
              description: Track daily pain levels and medication effectiveness
              questionnaireType: symptom_tracker
              frequency: daily
              isRequired: true
              estimatedTimeMinutes: 5
              questions:
                - id: pain_level
                  type: scale
                  text: How would you rate your pain level today?
                  required: true
                  validation:
                    min: 0
                    max: 10
                - id: medication_taken
                  type: boolean
                  text: Did you take your prescribed medication today?
                  required: true
      responses:
        '201':
          description: Questionnaire created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Questionnaire created successfully
                  questionnaire:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      title:
                        type: string
                      version:
                        type: string
                      status:
                        type: string
                        example: draft
                      createdAt:
                        type: string
                        format: date-time
        '403':
          description: Physician access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # Response Management Endpoints
...